Put Get_Locations() into a new ajax method. Added some logic to see if settings have changed.
Get_Locations() was never called (it was using old Laravel code to call it, which doesn't exist).
Now it is called properly, but ONLY when relevant data has changed for the address. Since
the class was never instantiated before, this is a step in the right direction and adding
logic for settings change for the other fields is outside scope of this ticket.

Get_Locations() is instantiated via event scheduling so it can be run in the background.