make idx_save_lead_settings and idx_put_lead static
in PHP8 it is a fatal error to call non-static methods in a static way.

this is something that CF7 does with the functions we provide in idx\leads\class-contact-form-7.php and most of the functions are static except for idx_save_lead_settings and idx_put_lead.

i don't see any issues with making these functions static as well and doing so makes our plugin compatible with cf7 in PHP8.