add impress-property-showcase to classname set for showcase widget
impress-property-showcase is the classname used by our CSS to style the widget and arrange the listings according to the "listings per row" option---this classname property is also used by page builders like beaver builder when inserting our widgets into the page.

prior to this change the classname set on the showcase widget was impress-showcase-widget which was not target by our css.

impress-showcase-widget has been left in the classname to avoid breaking any CSS created by clients or developer partners that might be using it to style our widget.