optimize updating of IDX pages
make a single query for get_posts when updating IDX pages with find_and_update_post to reduce number of calls made to database.

delete duplicate IDX pages that have an incorrect page slug.

decode page title from wordpress when checking to see if the title needs to be updated to match the link or page in the IDXB account