Mitigate OOM errors by maintaining a smaller array of properties to return from client_properties when possible
Reduce possibility of running into an OOM error by doing the following:

$max number of listings needed for the widget is used to limit the number of returned listings to that number

$user_agent_id is used to proactively filter listings out of the results

$max_sort is used together with $max to ensure the correct ordering of results is returned when low-high, high-low is set

Added 'none' as a sort order option for shortcodes if the order of the returned listings does not matter. Using none for the sort order lets the plugin return results much faster.

Note that functionality when 'default' is used for the sort order is kept the same to avoid disrupting the functionality of shortcode already being used on sites.